requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
chromadb
faiss-cpu
langchain
langchain_google_genai
langchain-community
langchain-chroma
pypdf
pysqlite3-binary